project/requirements-ci.txt
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
datasets==3.2.0
huggingface_hub==0.27.1
python-dotenv==1.0.1
